+++ src/PrestaShopBundle/Resources/config/services/core_services.yml
@@ -1,0 +1,2 @@
+imports:
+    - { resource: core/import.yml }
